Rename services to providers
