+++ llm_council/providers/cohere_provider.py
@@ -2,22 +2,24 @@ import dotenv
 import os
 import logging
 
-from llm_council.processors.services.base_service import BaseService
+from llm_council.providers.base_provider import BaseProvider
+from llm_council.providers.base_provider import provider
 
 dotenv.load_dotenv()
 
 
-class TogetherService(BaseService):
-    """https://docs.together.ai/reference/chat-completions"""
+@provider(provider_name="cohere", api_key_name="COHERE_API_KEY")
+class CohereProvider(BaseProvider):
+    """https://docs.cohere.com/reference/chat"""
 
     def __init__(self, llm) -> None:
-        BaseService.__init__(self, llm)
+        BaseProvider.__init__(self, llm)
 
     def __api_key(self) -> str | None:
-        return os.getenv("TOGETHER_API_KEY")
+        return os.getenv("COHERE_API_KEY")
 
     def request_url(self) -> str:
-        return "https://api.together.xyz/v1/chat/completions"
+        return "https://api.cohere.ai/v1/chat"
 
     def request_header(self) -> dict:
         return {
@@ -27,46 +29,47 @@ class TogetherService(BaseService):
 
     def sample_request(self) -> dict:
         return {
-            "model": "mistralai/Mistral-7B-Instruct-v0.2",
-            "messages": [{"role": "user", "content": "Say hello!"}],
+            "model": "command-r",
+            "message": "Say hello!",
         }
 
     def rate_limit_time_unit(self) -> str:
         return "seconds"
 
     def max_requests_per_unit(self) -> int:
-        return 9
+        return 95
 
     def get_request_prompt(self, request: dict) -> str:
-        return request["messages"][0]["content"]
+        return request["message"]
 
     def get_request_body(
         self, user_prompt: str, temperature: float | None, schema_name: str | None
     ) -> dict:
         if schema_name is not None:
             logging.warning(
-                f"Together does not support structured output. Skipping schema: {schema_name}."
+                f"Cohere does not support structured output. Skipping schema: {schema_name}."
             )
         if temperature is not None:
             return {
                 "model": self.model_name,
-                "messages": [{"role": "user", "content": user_prompt}],
+                "message": user_prompt,
                 "temperature": temperature,
             }
         else:
             return {
                 "model": self.model_name,
-                "messages": [{"role": "user", "content": user_prompt}],
+                "message": user_prompt,
             }
 
     def get_response_string(self, json_response: dict) -> str:
-        return json_response["choices"][0]["message"]["content"]
+        return json_response["text"]
 
     def get_response_info(self, json_response: dict) -> dict:
         return {
             "llm": self.llm,
             "model_name": self.model_name,
-            "id": json_response["id"],
-            "usage": json_response["usage"],
-            "model": json_response["model"],
+            "response_id": json_response["response_id"],
+            "generation_id": json_response["generation_id"],
+            "finish_reason": json_response["finish_reason"],
+            "meta": json_response["meta"],
         }
