+++ llm_council/providers/together_provider.py
@@ -2,22 +2,24 @@ import dotenv
 import os
 import logging
 
-from llm_council.processors.services.base_service import BaseService
+from llm_council.providers.base_provider import BaseProvider
+from llm_council.providers.base_provider import provider
 
 dotenv.load_dotenv()
 
 
-class CohereService(BaseService):
-    """https://docs.cohere.com/reference/chat"""
+@provider(provider_name="together", api_key_name="TOGETHER_API_KEY")
+class TogetherProvider(BaseProvider):
+    """https://docs.together.ai/reference/chat-completions"""
 
     def __init__(self, llm) -> None:
-        BaseService.__init__(self, llm)
+        BaseProvider.__init__(self, llm)
 
     def __api_key(self) -> str | None:
-        return os.getenv("COHERE_API_KEY")
+        return os.getenv("TOGETHER_API_KEY")
 
     def request_url(self) -> str:
-        return "https://api.cohere.ai/v1/chat"
+        return "https://api.together.xyz/v1/chat/completions"
 
     def request_header(self) -> dict:
         return {
@@ -27,47 +29,46 @@ class CohereService(BaseService):
 
     def sample_request(self) -> dict:
         return {
-            "model": "command-r",
-            "message": "Say hello!",
+            "model": "mistralai/Mistral-7B-Instruct-v0.2",
+            "messages": [{"role": "user", "content": "Say hello!"}],
         }
 
     def rate_limit_time_unit(self) -> str:
         return "seconds"
 
     def max_requests_per_unit(self) -> int:
-        return 95
+        return 9
 
     def get_request_prompt(self, request: dict) -> str:
-        return request["message"]
+        return request["messages"][0]["content"]
 
     def get_request_body(
         self, user_prompt: str, temperature: float | None, schema_name: str | None
     ) -> dict:
         if schema_name is not None:
             logging.warning(
-                f"Cohere does not support structured output. Skipping schema: {schema_name}."
+                f"Together does not support structured output. Skipping schema: {schema_name}."
             )
         if temperature is not None:
             return {
                 "model": self.model_name,
-                "message": user_prompt,
+                "messages": [{"role": "user", "content": user_prompt}],
                 "temperature": temperature,
             }
         else:
             return {
                 "model": self.model_name,
-                "message": user_prompt,
+                "messages": [{"role": "user", "content": user_prompt}],
             }
 
     def get_response_string(self, json_response: dict) -> str:
-        return json_response["text"]
+        return json_response["choices"][0]["message"]["content"]
 
     def get_response_info(self, json_response: dict) -> dict:
         return {
             "llm": self.llm,
             "model_name": self.model_name,
-            "response_id": json_response["response_id"],
-            "generation_id": json_response["generation_id"],
-            "finish_reason": json_response["finish_reason"],
-            "meta": json_response["meta"],
+            "id": json_response["id"],
+            "usage": json_response["usage"],
+            "model": json_response["model"],
         }
